requirements.txt
streamlit
pandas
scikit-learn
google-generativeai
rapidfuzz
asyncio
python-dotenv